producers/requirements.txt
kafka-python==2.0.2
requests==2.31.0
httpx==0.24.1
